Refactor MF provider into modular components
- Restructured the MF provider by breaking it down into multiple components including `MfHttpClient`, `MfRateLimiter`, `MfDataParser`, and `MfApiClient`.
- Updated the main `MfProvider` class to utilize these components for improved maintainability and clarity.
- Added an `__init__.py` file to the `mf` module for better organization and accessibility of components.
- Enhanced error handling and response parsing within the new structure.
